refactor(cf): relocate cf service signal configs out of list framework
